Minor module reorganization. See #16
